pyproject.toml
[tool.poetry]
name = "university pass system for aitu"
version = "0.1"
description = ""
authors = ["Sagat Akimbay"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.0"
django-allauth = "^65.7.0"
requests = "^2.32.3"
pillow = "^11.2.1"
python-dotenv = "^1.1.0"
pyopenssl = "^25.0.0"
werkzeug = "^3.1.3"
django-extensions = "^4.1"
django-select2 = "^8.4.0"
django-filter = "^25.1"
openpyxl = "^3.1.5"
celery = "^5.5.1"
redis = "^5.2.1"
django-redis = "^5.4.0"
psycopg2-binary = "^2.9.10"
pyngrok = "^7.2.5"
gunicorn = "^23.0.0"
uvicorn = "^0.29.0"
flower = "^2.0.1"
django-pwa = "^2.0.1"
django-widget-tweaks = "^1.5.0"
faker = "^37.3.0"
whitenoise = "^6.9.0"
django-debug-toolbar = "^5.2.0"
qrcode = "^8.2"
pytz = "^2025.2"
sentry-sdk = {extras = ["django"], version = "^2.35.0"}
django-prometheus = "^2.3.0"
django-csp = "^3.8"
cryptography = "^43.0.0"
djangorestframework = "^3.15.2"
django-axes = "^6.5.0"
channels = "^4.0.0"
channels-redis = "^4.2.0"
drf-spectacular = "^0.27.2"
django-htmx = "^1.23.2"
django-guardian = "^3.1.0"
django-webpush = "^0.3.6"
pywebpush = "^2.0.3"
py-vapid = "^1.9.2"
requests-toolbelt = "^1.0.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
